requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiomysql==0.1.1
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.9.3
APScheduler==3.10.0
async-generator==1.10
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
decorator==5.1.1
environs==9.5.0
exceptiongroup==1.1.0
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
magic-filter==1.0.9
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
multidict==6.0.4
outcome==1.2.0
packaging==23.0
pymemcache==4.0.0
PyMySQL==1.0.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.44
SQLAlchemy-Ext==0.2
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
webdriver-manager==3.8.5
wsproto==1.2.0
yarl==1.8.2
zipp==3.12.1
